Problem: in C4, no way to divide bigger issues to smaller minimal patches
Solution: point C4 to 20-C4, which implements @dulanov's hierarchical
issues idea.

C4 is good about that each commit should be a minimal identifiable
problem, but does not have any machnisms inside to reach this
goal. 20-C4 tries to be a battery-included modification for C4, which
are being actively tested in this repo (like #122).